feat: Complete UWB integration and wheel loader system enhancements
🚀 Major Features Added:

## UWB (Ultra-Wideband) Integration for EKF2
- ✅ Added NoopLoop LinkTrack UWB driver with full PX4 integration
- ✅ Implemented complete UWB fusion pipeline in EKF2
- ✅ Added UWB range measurements for indoor positioning
- ✅ Created comprehensive UWB parameter system (16 parameters)
- ✅ Added UWB quality checks and signal validation
- ✅ Integrated UWB data flow: Driver → sensor_uwb → EKF2 fusion

### UWB Driver Features:
- Split architecture: .hpp/.cpp files following PX4 conventions
- LinkTrack protocol support with enhanced diagnostics
- Signal quality metrics (RSSI, LOS confidence, multipath detection)
- Configurable anchor positions and sensor mounting
- Robust error handling and parameter validation

### EKF2 UWB Fusion:
- Added UWB aid source with innovation gating
- Range-based position estimation with anchor mapping
- Adaptive measurement noise based on signal quality
- Comprehensive parameter system for tuning
- Build system integration with CONFIG_EKF2_UWB

## Wheel Loader System Enhancements:
- 🔧 Enhanced articulated chassis control systems
- 🔧 Improved traction control and slip estimation
- 🔧 Advanced terrain adaptation algorithms
- 🔧 Load-aware torque distribution
- 🔧 Updated message definitions for load sensing

## Build System & Configuration:
- ✅ Added UWB Kconfig options for proper feature selection
- ✅ Updated board configuration (hkust_nxt-dual-wl-rear) with UWB support
- ✅ Fixed parameter naming conventions (≤16 chars)
- ✅ Corrected parameter units following PX4 standards
- ✅ Added ekf2_timestamps logging for UWB debugging

## Technical Improvements:
- Fixed double promotion warnings in UWB signal processing
- Implemented proper aid source status management
- Added UWB buffer system with time synchronization
- Created comprehensive parameter validation
- Enhanced CMakeLists.txt structure for conditional compilation

📊 Validation Status:
- ✅ Successful compilation on ARM Cortex-M7 target
- ✅ Memory usage: 54.49% Flash, 10.90% SRAM
- ✅ All UWB files compiled and linked successfully
- ✅ Firmware generation completed: hkust_nxt-dual-wl-rear_default.px4

🔗 Integration Pipeline:
NoopLoop Driver → sensor_uwb topic → EKF2::UpdateUwbSample() → setUwbData() → UWB Buffer → controlUwbFusion() → Position Estimation

Ready for runtime testing and validation! 🎯

+++ src/modules/articulated_chassis/wheel_controller/wheel_controller.cpp
@@ -484,8 +484,8 @@ int WheelController::print_status()
     PX4_INFO("  Performance:");
     PX4_INFO("    Speed Error RMS: %.2f RPM", (double)_performance.speed_error_rms);
     PX4_INFO("    Control Effort: %.1f%%", (double)(_performance.control_effort_avg * 100.0f));
-    PX4_INFO("    Slip Events: %lu", _performance.slip_events);
-    PX4_INFO("    Safety Violations: %lu", _performance.safety_violations);
+    PX4_INFO("    Slip Events: %lu", (unsigned long)_performance.slip_events);
+    PX4_INFO("    Safety Violations: %lu", (unsigned long)_performance.safety_violations);
 
     perf_print_counter(_loop_perf);
     perf_print_counter(_control_latency_perf);
